checking basics

+++ CampGround/routes/campgrounds.js
@@ -75,23 +75,39 @@ router.post("/", isLoggedIn, validateCampground, async (req, res, next) => {
 });
 
 router.get("/:id/edit", isLoggedIn, async (req, res) => {
-    const campground = await Campground.findById(req.params.id);
+    const { id } = req.params;
+    const campground = await Campground.findById(id);
     if(!campground){
         req.flash("error","Can't find the campground");
         return res.redirect("/campgrounds");
+    }
+    if(!campground.author.equals(req.user._id)){
+        req.flash("error", "You can't do that, Access Denied");
+        return res.redirect(`/campgrounds/${campground._id}`)
     }
     res.render("campgrounds/edit.ejs", { campground });
 });
 
 router.put("/:id", isLoggedIn, validateCampground, async (req, res) => {
     const { id } = req.params;
-    const campground = await Campground.findByIdAndUpdate(id, { ...req.body.campground }, { new: true });
+    const campground = await Campground.findById(id)
+    if(!campground.author.equals(req.user._id)){
+        req.flash("error", "You can't do that, Access Denied");
+        return res.redirect(`/campgrounds/${campground._id}`)
+    }
+    //well/...
+    const camp = await Campground.findByIdAndUpdate(id, { ...req.body.campground }, { new: true });
     req.flash("success","Successfully updated")
     res.redirect(`/campgrounds/${campground._id}`);
 });
 
 router.delete("/:id", isLoggedIn, async (req, res) => {
     const { id } = req.params;
+    const campground = await Campground.findById(id)
+    if(!campground.author.equals(req.user._id)){
+        req.flash("error", "You can't do that, Access Denied");
+        return res.redirect(`/campgrounds/${campground._id}`)
+    }
     await Campground.findByIdAndDelete(id);
     req.flash("success","Successfully deleted camp");
     res.redirect("/campgrounds");
